refactor(G1): extract resolution chain into resolver/resolve.rs
resolver/mod.rs is now a 48-line thin coordinator (module declarations
+ re-exports only) — matching the rust-analyzer lsp.rs/cli.rs pattern.

The full resolution chain (ensure_file_data, fqns_for_name,
resolve_symbol, resolve_symbol_inner, resolve_symbol_no_rg, all step
implementations, rg_in_package_dir, is_stdlib, impl Indexer wrappers)
moves to resolver/resolve.rs (~700 lines, single responsibility).

No behaviour changes. 830 tests pass, 0 new clippy warnings.
